services/gpu-worker/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML dependencies
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
diffusers==0.24.0
transformers==4.36.0
accelerate==0.25.0
safetensors==0.4.1
compel==2.0.2

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# Utilities
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2
psutil==5.9.6
GPUtil==1.4.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Monitoring and logging
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
